quantum_fraud_detection/utils/quantum_utils.py
from dimod import BinaryQuadraticModel
from dwave.system import DWaveSampler, EmbeddingComposite
from dwave.samplers import SimulatedAnnealingSampler

def quantum_inspired_optimization(graph_data):
    # Simulate quantum annealing for graph partitioning (fraud cascade detection)
    # Simplified: Model as QUBO for min-cut (NP-hard)
    bqm = BinaryQuadraticModel('BINARY')
    for node in graph_data.nodes():
        bqm.add_variable(node, 1.0)  # Bias for partitioning
    for u, v in graph_data.edges():
        bqm.add_interaction(u, v, -2.0)  # Coupler
    
    # Use simulated annealing (no real quantum hardware)
    sampler = SimulatedAnnealingSampler()
    response = sampler.sample(bqm, num_reads=10)
    best_sample = response.first.sample
    # Extract partitions (0/1 for legit/fraud clusters)
    partitions = [node for node, val in best_sample.items() if val == 1]
    return partitions